[EM-11] package by domain refactor
