+++ src/main/java/it/andmora/expensesmonitor/dao/persistance/PaymentRepositorySample.java
@@ -3,6 +3,6 @@ package it.andmora.expensesmonitor.dao.persistance;
 /**
  * The persistance layer implements the repository or anything needed to persist the domain objects
  */
-public interface PaymentRepositorySample {
+interface PaymentRepositorySample {
 
 }
