+++ src/main/java/it/andmora/expensesmonitor/domain/usecase/PaymentCreatorImpl.java
@@ -1,11 +1,11 @@
 package it.andmora.expensesmonitor.domain.usecase;
 
-import it.andmora.expensesmonitor.dao.PaymentDao;
+import it.andmora.expensesmonitor.domain.PaymentDao;
 import it.andmora.expensesmonitor.domain.entity.Payment;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class PaymentCreatorImpl implements PaymentCreator {
+class PaymentCreatorImpl implements PaymentCreator {
 
   private final PaymentDao paymentDao;
